requirements.txt
python-telegram-bot[ext]==21.6
openai>=1.40.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
